requirements.txt
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0 
yt-dlp
ffmpeg-python
openai-whisper
torch
textblob
langdetect
google-cloud-videointelligence
certifi 
streamlit 
groq
openai>=1.0.0
openrouter>=0.1.3  # wrapper for OpenRouter.ai compatible with openai-python
python-dotenv>=0.19.0  # Load environment variables from .env file
requests>=2.25.0  # HTTP library for API calls
pillow 